python/requirements.txt
langchain>=0.1.14,<0.2
langchain-community>=0.0.30,<0.1
langchain-openai>=0.1.3,<0.2
python-dotenv>=1.0.0,<2.0
SQLAlchemy>=2.0.25,<3.0
pymysql>=1.1.0,<2.0
psycopg2-binary>=2.9.0,<3.0